Added count to nav and seperated reducer and store
