+++ redux/count/selector.js
@@ -1,0 +1,1 @@
+export const selectWishlistCount = (state) => state.count.count;
