+++ redux/store.js
@@ -1,7 +1,7 @@
-import { createStore, applyMiddleware, combineReducers } from "redux";
-import { HYDRATE, createWrapper } from "next-redux-wrapper";
+import { createStore, applyMiddleware } from "redux";
+import { createWrapper } from "next-redux-wrapper";
 import thunkMiddleware from "redux-thunk";
-import count from "./count/reducer";
+import reducer from "./reducer";
 
 const bindMiddleware = (middleware) => {
   if (process.env.NODE_ENV !== "production") {
@@ -9,23 +9,6 @@ const bindMiddleware = (middleware) => {
     return composeWithDevTools(applyMiddleware(...middleware));
   }
   return applyMiddleware(...middleware);
-};
-
-const combinedReducer = combineReducers({
-  count,
-});
-
-const reducer = (state, action) => {
-  if (action.type === HYDRATE) {
-    const nextState = {
-      ...state, // use previous state
-      ...action.payload, // apply delta from hydration
-    };
-    if (state.count) nextState.count = state.count; // preserve count value on client side navigation
-    return nextState;
-  } else {
-    return combinedReducer(state, action);
-  }
 };
 
 const initStore = () => {
